+++ src/test/java/org/tinyradius/core/attribute/type/OctetsAttributeTest.java
@@ -20,7 +20,7 @@ class OctetsAttributeTest {
     @Test
     void createMaxSizeAttribute() {
         // 253 octets ok
-        final RadiusAttribute maxSizeAttribute = new OctetsAttribute(dictionary, -1, (byte) 2, random.generateSeed(253));
+        final RadiusAttribute maxSizeAttribute = new OctetsAttribute(dictionary, -1, 2, random.generateSeed(253));
         final byte[] bytes = maxSizeAttribute.toByteArray();
 
         assertEquals(0xFF, toUnsignedInt(bytes[1]));
@@ -30,20 +30,20 @@ class OctetsAttributeTest {
         // 254 octets not ok
         final byte[] oversizedArray = random.generateSeed(254);
         final IllegalArgumentException exception = assertThrows(IllegalArgumentException.class, () ->
-                new OctetsAttribute(dictionary, -1, (byte) 2, oversizedArray));
+                new OctetsAttribute(dictionary, -1, 2, oversizedArray));
 
         assertTrue(exception.getMessage().contains("too long"));
     }
 
     @Test
     void testFlatten() {
-        final RadiusAttribute attribute = new OctetsAttribute(dictionary, -1, (byte) 2, "123456");
+        final RadiusAttribute attribute = new OctetsAttribute(dictionary, -1, 2, "123456");
         assertEquals("[User-Password: 123456]", attribute.flatten().toString());
     }
 
     @Test
     void testToString() {
-        final RadiusAttribute attribute = new OctetsAttribute(dictionary, -1, (byte) 2, "123456");
+        final RadiusAttribute attribute = new OctetsAttribute(dictionary, -1, 2, "123456");
         assertEquals("User-Password: 123456", attribute.toString());
     }
 
@@ -54,7 +54,7 @@ class OctetsAttributeTest {
 
         final byte[] requestAuth = random.generateSeed(16);
 
-        final RadiusAttribute attribute = dictionary.createAttribute(-1, (byte) 2, (byte) 0, pw);
+        final RadiusAttribute attribute = dictionary.createAttribute(-1, 2, (byte) 0, pw);
         assertTrue(attribute instanceof OctetsAttribute);
         assertFalse(attribute.isEncoded());
         assertEquals(pw, new String(attribute.getValue(), UTF_8));
