+++ src/test/java/org/tinyradius/core/attribute/type/StringAttributeTest.java
@@ -16,7 +16,7 @@ class StringAttributeTest {
     @Test
     void dataBadSizes() {
         final IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> new StringAttribute(dictionary, -1, (byte) 1, ""));
+                () -> new StringAttribute(dictionary, -1, 1, ""));
 
         assertTrue(exception.getMessage().toLowerCase().contains("min 1 octet"));
     }
@@ -24,7 +24,7 @@ class StringAttributeTest {
     @Test
     void getDataValue() {
         final String s = new Date().toString();
-        final StringAttribute stringAttribute = new StringAttribute(dictionary, -1, (byte) 1, s);
+        final StringAttribute stringAttribute = new StringAttribute(dictionary, -1, 1, s);
 
         assertEquals(s, stringAttribute.getValueString());
         assertArrayEquals(s.getBytes(UTF_8), stringAttribute.getValue());
