+++ src/test/java/org/tinyradius/core/packet/request/AccountingRequestTest.java
@@ -2,11 +2,11 @@ package org.tinyradius.core.packet.request;
 
 import net.jradius.util.RadiusUtils;
 import org.junit.jupiter.api.Test;
+import org.tinyradius.core.RadiusPacketException;
 import org.tinyradius.core.attribute.type.IntegerAttribute;
 import org.tinyradius.core.dictionary.DefaultDictionary;
 import org.tinyradius.core.dictionary.Dictionary;
 import org.tinyradius.core.packet.PacketType;
-import org.tinyradius.core.RadiusPacketException;
 
 import java.util.Collections;
 
@@ -26,7 +26,7 @@ class AccountingRequestTest {
         String sharedSecret = "sharedSecret";
         String user = "myUser1";
         RadiusRequest request = new AccountingRequest(dictionary, (byte) 1, null, Collections.emptyList())
-                .addAttribute(dictionary.createAttribute(-1, (byte) 1, user.getBytes(UTF_8)))
+                .addAttribute(dictionary.createAttribute(-1, 1, user.getBytes(UTF_8)))
                 .addAttribute("Acct-Status-Type", "7");
 
         final byte[] attributeBytes = request.getAttributeBytes();
